Remove Brewfile.lock.json
The lockfile changes almost everytime running 'make install' (which
contains 'brew bundle dump') so the change history easily gets messed
up.
Plus, Brewfile.lock.json is not used to fix versions for later installation but just
to record versions at the installation time. Thus it is not necessary to store
it.